requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
psycopg2-binary
langchain
langchain-google-genai
python-dotenv
authlib
psycopg2-binary
sqlalchemy